Add tests for PackService and code refactoring
1)Refactored ViewController's endpoint that returns the packs view to do the
conversion from Pack entities to DTOs instead of doing it internally inside PackService.
This is done to be consistent with the rest of the services that return entities. Also, it
moves the responsibility of conversion from service layer to controller layer.
2)Refactored Pack entity - it no longer accepts region as constructor parameter and instead passes the region
of the creation location.
3)Refactored method in ItemPriceRepository - it now accepts items instead of ids.
4)Refactored method in PackRepository - used left join where possible to make testing easier since
without left join fetch, tests would have to create full pack entities that contain recipes and materials.
Also, this method accepts additional parameter which specified the continent of the pack, and it's used in the
where clause of the query.
5)Added ItemPriceService - used by ViewController to get prices of items.
6)Refactored PackService:
 a)packs method now returns entities.
 b)the continent parameter is used to filter out returned packs.
7)Added static method to AbstractTest that is used as method source for parameterized tests in PackServiceITest.
8)Added repositories to AbstractITest so that subclasses can easily use them without specifying them as dependencies.
9)Added PackServiceITest that contains integration tests for PackService.
10)Change logging levels for some loggers in application-test.properties so that we can see the sql queries and
bind parameters in the logs when executing tests.