Refactor PackService and integration tests
1)Refactored ViewController to no longer convert returned packs from PackService to PackDTO. Instead, this code
is moved inside PackService and is part of the business logic. This is done because conversion and sorting of packs
by profit is part of the business logic and therefore should be included in service layer.
Also, view no longer adds prices attribute, and instead adds materials attribute that contains materials which have price.
2)Added init constructs in ItemPrice and Recipe that establishes the bidirectional association. This is done to
simplify tests which did not work correctly without establishing proper bidirectional association.
3)Updated PackRepository not to include region in the where clause - it's not needed because we already filter by using
the archeage server in the where clause which belongs to specific region.
4)Updated PackService to return PackDTOs instead of Packs.
5)Added extension method to utils that converts packs to DTOs - the reason for it being in utils and not in PackService(the place
where it's used), is to be able to use it in tests as well.
6)Added dev context to update changesets that update the sequence generator tables. Without it, tests were using negative ids.
7)Updated packs.html to use the new materials attribute instead of prices.
8)Updated PackService and added one more test case for method that accepts departure location. TODO more tests for other
methods.