Add integration tests for ItemPriceService and fix bugs
1)Refactored PurchasableItem: made prices private property and added two methods - one adds a new price
and the second returns all prices wrapped in unmodifiable wrapper. This is done to prevent errors where
invalid prices can be added.
2)Refactored UserPrice: added init block that checks if the user price object is valid and adds itself
to the item's prices collection.
3)Refactored method in PackRecipeRepository that returned crafting materials to only return crafting
materials that are purchasable since we use them in profile view to allow user to change their price which
for non-purchasable items is not possible.
4)Refactored UserPriceRepository to also include additional expression in the where clause of sql queries
which is to only return prices of specific archeage server. This is done since we want to use only prices
that belong to the user's currently chosen archeage server.
5)Fixed prices.js which used first item for all prices instead of using the iteration variable.
6)Added integration tests for ItemPriceService.