Improve pack pagination to return packs sorted
1)Added profit method to Pack entity that returns the profit of the pack given material prices.
This method is used in the new PackProfitService to calculate the initial pack profit.
2)Refactored PackProfit entity:
 a)Added default value for timestamp.
 b)Renamed goldProfit property from goldProfit into netProfit since the goldProfit doesn't
 store profit of gold, but the price in general.
3)Added new spring event class - ItemPriceChangeEvent that is fired by ItemPriceService when
new price is created for an item.
4)Added PackProfitRepository with methods that previous were in PackRepository - methods that
return pack ids. The reason for moving them from PackRepository to PackProfitRepository was that
if pack profit exists for a pack, it means that all price materials have been specified by some user
for the pack.
5)Removed methods from PackRepository as explained in (4).Also , added one more method that returns
packs for which user has specified prices for all purchasable items that are used as crafting materials.
6)Refactored ItemPriceServiceImpl to add ItemPriceChangeEvent publication.
7)Added PackProfitService with single method that adds/updates pack profits when ItemPriceChangeEvent
is fired.
8)Refactored PackServiceImpl to use PackProfitRepository methods instead of the removed PackRepository
methods.
Also,removed sorting by profit because sorting can now be specified in Pageable request and done by
database.
9)Removed unused method from utils - materialsWithPrice.
10)Refactored packs.js to add sort query parameters to request packs sorted by net profit.
11)Update tests due to changes.