Add tests for packs page when user authenticated
1)Removed AbstractAuthenticatedPageObject and AbstractUnauthenticatedPageObject because currently it does not provide any value.
2)Added AuthenticationData data class used by LoginPageObject to perform authentication. Additionally to email and password,
it also provides user id which is needed by AuthenticatedPacksPageObject.
3)Renamed Packages* test classes to Packs*.
4)Made PacksPageObject open so that AuthenticatedPacksPageObject and extend it. Additionally, refactored changePercentage method to
wait for new packs.
Finally, added userID() method that returns the id of the user that is used to retrieve current packs. By default, it returns null because
PacksPageObject assumes unauthenticated scenario, but this method can be overridden for scenarios when user is authenticated.
5)Added AuthenticatedPacksPageObject that extends PacksPageObject and performs authentication on load.
6)Added AuthenticatedPacksPageTest that contains tests for packs page when user is authenticated.
7)Refactored AbstractPacksPageTest not to define page field and instead subclasses define their own page objects. This is done because AuthenticatedPacksPageTest
extends it and has to define different type of page object - AuthenticatedPacksPageObject.